Temporary bug fix - variable stars
If stars were variable, a value in the "Variable?" column in 'Star Data.txt' would be equal to 1 (as opposed to 0). However, the photometry data would only be output if a star was variable AND was a certain distance away. For stars that are a certain distance away but still variable, a bug would be encountered while automatically processing data in analysis (NOT during simulation).